+++ src/deploy/actions.spec.ts
@@ -1,9 +1,9 @@
 import { JsonObject, logging } from '@angular-devkit/core';
 import { BuilderContext, BuilderRun, ScheduleOptions, Target } from '@angular-devkit/architect/src/index';
 import deploy from './actions';
+import { SynchronousDelegateExpectedException } from '@angular-devkit/core/src/virtual-fs/host';
 
 let context: BuilderContext;
-const mockEngine = { run: (_: string, __: any, __2: any) => Promise.resolve() }
 
 const PROJECT = 'pirojok-project';
 
@@ -12,7 +12,7 @@ describe('Deploy Angular apps', () => {
 
   it('should invoke the builder', async () => {
     const spy = spyOn(context, 'scheduleTarget').and.callThrough();
-    await deploy(mockEngine, context, 'host', {});
+    await deploy( context, 'host',"app", {});
 
     expect(spy).toHaveBeenCalledWith({
         target: 'build',
@@ -25,7 +25,7 @@ describe('Deploy Angular apps', () => {
 
   it('should invoke the builder with the baseHref', async () => {
     const spy = spyOn(context, 'scheduleTarget').and.callThrough();
-    await deploy(mockEngine, context, 'host', { baseHref: '/folder'});
+    await deploy(context, 'host',"app", { baseHref: '/folder'});
 
     expect(spy).toHaveBeenCalledWith({
         target: 'build',
@@ -36,22 +36,34 @@ describe('Deploy Angular apps', () => {
     );
   });
 
-  it('should invoke engine.run', async () => {
-    const spy = spyOn(mockEngine, 'run').and.callThrough();
-    await deploy(mockEngine, context, 'host', {});
+  // it('should invoke engine.run', async () => {
+  //   const spy = spyOn(mockEngine, 'run').and.callThrough();
+  //   await deploy(mockEngine, context, 'host', {});
 
-    expect(spy).toHaveBeenCalledWith('host', {}, context.logger);
-  });
+  //   expect(spy).toHaveBeenCalledWith('host', {}, context.logger);
+  // });
 
   describe('error handling', () => {
     it('throws if there is no target project', async () => {
       context.target = undefined;
       try {
-        await deploy(mockEngine, context, 'host', {});
+        await deploy(context, 'host', "app", {});
         fail();
       } catch (e) {
         expect(e.message).toMatch(/Cannot execute the build target/);
       }
+    });
+
+    it('return false if api Token not present', async () => {
+      var output = await deploy(context, 'host', "app", {});
+      expect(output.success).toEqual(false);
+    });
+    
+    it('return false if app name not present', async () => {
+      var output = await deploy(context, 'host', "app", {
+        herokuApiToken:"asd"
+      });
+      expect(output.success).toEqual(false);
     });
   });
 });
