+++ src/engine/engine.spec.ts
@@ -1,10 +1,23 @@
 import * as engine from './engine';
-import { NullLogger } from '@angular-devkit/core/src/logger';
+import { JsonObject, logging } from '@angular-devkit/core';
+import { Logger, LogLevel } from '@angular-devkit/core/src/logger';
+const Heroku = require('heroku-client');
+const logger: any = {
+  error: (message: string) => { }
+}
+let heroku: any = null;
+jest.mock('heroku-client') // this auto mocks all methods on heroku-client
 
 describe('engine', () => {
-  it('should copy directory', () => {
-
-    // TODO: really write test here!
-    expect(1 + 1).toEqual(2);
+  beforeEach(() => {
+    heroku = new Heroku({ token: 'asd' });
   });
+
+  it('should return false if application name not find in Heroku', async () => {
+    const spy = spyOn(logger, 'error').and.callThrough();
+    const result = await engine.run("", { appName: 'test' }, "app", logger);
+    expect(result.success).toEqual(false);
+    expect(spy).toHaveBeenCalledWith("🚨 test application not found in Heroku!");
+  });
+
 });
