--- src/ng-add-schema.json
@@ -1,14 +1,0 @@
-{
-  "$schema": "http://json-schema.org/schema",
-  "id": "angular-cli-ghpages-ng-add-schematic",
-  "title": "angular-cli-ghpages ng-add schematic",
-  "type": "object",
-  "properties": {
-    "project": {
-      "type": "string",
-      "description": "The name of the project."
-    }
-  },
-  "required": [],
-  "additionalProperties": false
-}
