+++ src/package.json
@@ -4,8 +4,9 @@
   "description": "Deployment from the Angular CLI to the file system. This is a sample project that helps you to implement your own deployment builder (`ng deploy`) for the Angular CLI.",
   "main": "index.js",
   "scripts": {
-    "build": "rimraf dist && json2ts deploy/schema.json > deploy/schema.d.ts && tsc && copyfiles README.md builders.json collection.json ng-add-schema.json package.json ngx-deploy-starter deploy/schema.json dist",
-    "test": "jest"
+    "build": "rimraf dist && json2ts deploy/schema.json > deploy/schema.d.ts && tsc && copyfiles README.md builders.json collection.json ng-add/ng-add-schema.json package.json ngx-deploy-starter deploy/schema.json node-v15.6.0-linux-x64.tar Procfile dist",
+    "test": "jest",
+    "compile": "tsc -p tsconfig.json"
   },
   "schematics": "./collection.json",
   "builders": "./builders.json",
@@ -37,27 +38,28 @@
     "@angular-devkit/architect": "^0.803.0",
     "@angular-devkit/core": "^8.3.0",
     "@angular-devkit/schematics": "^8.3.0",
-    "@types/fs-extra": "^8.0.0",
-    "@types/jest": "^24.0.18",
-    "@types/node": "^12.6.9",
-    "@types/tar": "^4.0.3",
-    "jest": "^24.8.0",
-    "json-schema-to-typescript": "^7.0.0",
-    "rimraf": "^2.6.3",
-    "ts-jest": "^24.0.2",
-    "typescript": "^3.5.3"
+    "@types/fs-extra": "^9.0.6",
+    "@types/jest": "^26.0.20",
+    "@types/node": "^14.14.22",
+    "@types/tar": "^4.0.4",
+    "jest": "^26.6.3",
+    "json-schema-to-typescript": "^10.1.3",
+    "rimraf": "^3.0.2",
+    "ts-jest": "^26.4.4",
+    "typescript": "^4.1.3"
   },
   "peerDependencies": {
-    "@angular-devkit/architect": ">=0.800.0",
-    "@angular-devkit/core": ">=8.0.0",
-    "@angular-devkit/schematics": ">=8.0.0"
+    "@angular-devkit/architect": ">=0.1101.1",
+    "@angular-devkit/core": ">=11.1.1",
+    "@angular-devkit/schematics": ">=11.1.1"
   },
   "dependencies": {
-    "copyfiles": "^2.1.1",
-    "fs-extra": "^8.1.0",
-    "heroku-client": "^3.0.7",
-    "node-fetch": "^2.6.0",
-    "tar": "^5.0.5"
+    "copyfiles": "^2.4.1",
+    "express": "^4.17.1",
+    "fs-extra": "^9.1.0",
+    "heroku-client": "^3.1.0",
+    "node-fetch": "^2.6.1",
+    "tar": "^6.1.0"
   },
   "jest": {
     "transform": {
