+++ src/public_api.ts
@@ -1,3 +1,3 @@
-export * from './ng-add';
+export * from './ng-add/ng-add';
 export * from './deploy/actions';
 export * from './deploy/builder';
