+++ src/tsconfig.json
@@ -24,7 +24,7 @@
     "moduleResolution": "node",
     "esModuleInterop": true
   },
-  "files": ["index.ts"],
+  "files": ["index.ts","server.ts"],
   "angularCompilerOptions": {
     "skipTemplateCodegen": true,
     "strictMetadataEmit": true,
